# fix: replication stuck when syncing large batches