--- a/crates/fluvio-spu/src/replication/test.rs
+++ b/crates/fluvio-spu/src/replication/test.rs
@@ -3,9 +3,12 @@ use std::sync::Arc;
 use std::time::Duration;
 use std::env::temp_dir;
 
+use fluvio::Isolation;
 use fluvio_controlplane::replica::Replica;
 use fluvio_controlplane::spu_api::update_replica::UpdateReplicaRequest;
+use fluvio_storage::iterators::FileBatchIterator;
 use fluvio_types::event::offsets::OffsetPublisher;
+use rand::Rng;
 use tracing::debug;
 use derive_builder::Builder;
 use once_cell::sync::Lazy;
@@ -874,7 +877,9 @@ async fn test_sync_2_replicas_but_one_reject() {
 
 #[fluvio_future::test(ignore)]
 async fn test_sync_larger_records() {
-    let num_records = 5;
+    let num_records_total = 100;
+    let num_records_per_batch = 10;
+    let num_batches = 10;
     let port = portpicker::pick_unused_port().expect("No free ports left");
     let builder = TestConfig::builder()
         .followers(1_u16)
@@ -910,18 +915,30 @@ async fn test_sync_larger_records() {
     assert_eq!(leader.hw(), 0);
 
     // create a raw record with 512kb
-    let record_test: Vec<u8> = vec![10; 512 * 1024];
+    let mut rng = rand::thread_rng();
+    let records = Vec::from_iter(
+        (0..num_batches)
+            .map(|_| {
+                let mut record = vec![0; 512 * 1024];
+                rng.fill(&mut record[..]);
+                record
+            })
+            .collect::<Vec<Vec<u8>>>()
+            .into_iter(),
+    );
 
-    // write a batch with 5 records with 512kb, total of 2.5MB
-    leader
-        .write_record_set(
-            &mut create_raw_recordset_inner(num_records, &record_test),
-            leader_gctx.follower_notifier(),
-        )
-        .await
-        .expect("write");
+    // write 10 batches with 10 records with 512kb, total of 5MB per batch and 50MB total
+    for record in &records {
+        leader
+            .write_record_set(
+                &mut create_raw_recordset_inner(num_records_per_batch, record),
+                leader_gctx.follower_notifier(),
+            )
+            .await
+            .expect("write");
+    }
 
-    assert_eq!(leader.leo(), num_records as i64);
+    assert_eq!(leader.leo(), num_records_total as i64);
     assert_eq!(leader.hw(), 0);
 
     let follower_gctx = builder.follower_ctx(0).await;
@@ -940,11 +957,65 @@ async fn test_sync_larger_records() {
 
     // wait until follower sync up with leader
     sleep(Duration::from_millis(*MAX_WAIT_REPLICATION)).await;
-    assert_eq!(follower.leo(), num_records as i64);
+    assert_eq!(follower.leo(), num_records_total as i64);
 
     // hw has been replicated
-    assert_eq!(follower.hw(), num_records as i64);
-    assert_eq!(leader.hw(), num_records as i64);
+    assert_eq!(follower.hw(), num_records_total as i64);
+    assert_eq!(leader.hw(), num_records_total as i64);
+
+    // check if the records are the same
+    let leader_replica = leader
+        .read_records(0, num_records_total as u32, Isolation::ReadCommitted)
+        .await
+        .expect("read leader records");
+    let follower_replica = follower
+        .read_records(0, num_records_total as u32, Isolation::ReadCommitted)
+        .await
+        .expect("read follower records");
+
+    assert_eq!(leader_replica.start, follower_replica.start);
+    assert_eq!(leader_replica.end, follower_replica.end);
+    let leader_slice = leader_replica.file_slice.expect("slice");
+    let follower_slice = follower_replica.file_slice.expect("slice");
+
+    assert_eq!(leader_slice.len(), num_records_total as u64);
+    assert_eq!(follower_slice.len(), num_records_total as u64);
+
+    follower_gctx
+        .replica_localstore()
+        .sync_all(vec![replica.clone()]);
+
+    let mut batch_leader = FileBatchIterator::from_raw_slice(leader_slice);
+    let mut batch_follower = FileBatchIterator::from_raw_slice(follower_slice);
+
+    let mut leader_batches = vec![];
+    let mut follower_batches = vec![];
+    while let Some(Ok(record)) = batch_leader.next() {
+        leader_batches.push(record);
+    }
+    while let Some(Ok(record)) = batch_follower.next() {
+        follower_batches.push(record);
+    }
+    assert_eq!(leader_batches.len(), 1);
+    assert_eq!(follower_batches.len(), 1);
+
+    assert_eq!(
+        leader_batches[0].batch.base_offset,
+        follower_batches[0].batch.base_offset
+    );
+    assert_eq!(
+        leader_batches[0].batch.batch_len,
+        follower_batches[0].batch.batch_len
+    );
+
+    assert_eq!(leader_batches[0].records, follower_batches[0].records);
+    for (record_leader, record_follower) in leader_batches[0]
+        .records
+        .iter()
+        .zip(follower_batches[0].records.iter())
+    {
+        assert_eq!(record_leader, record_follower);
+    }
 
     sleep(Duration::from_millis(WAIT_TERMINATE)).await;