Should fix replica sync problem: https://github.com/infinyon/fluvio/issues/4472


I also notice more perfomance with replicas now:

before:
```
 fluvio benchmark producer -p 3 -r 3 -k --num-records 50000                                                                                                                                                                                                                                                              $ fix_replication_large_2 
Benchmark started

latencies: 19.7µs min, 7.4ms avg, 33.1ms max, 7.2ms p0.50, 15.1ms p0.95, 17.3ms p0.99
50000 total records sent, 71942 records/sec: (369.1 MB/sec), total time: 695.9ms

**Per Record E2E Latency**

|Variable| p0.00 |p0.50|p0.95 |p0.99 |p1.00 |
|--------|-------|-----|------|------|------|
|Latency |19.7µs |7.2ms|15.1ms|17.3ms|33.1ms|

**Throughput (Total Produced Bytes / Time)**

|     Variable      |   Speed    |
|-------------------|------------|
|Produced Throughput|369.1 MB/sec|

Benchmark completed
```


now: 

```
▷ flvdr benchmark producer -p 3 -r 3 -k --num-records 50000                                                                                                                                                                                                                                                               $ fix_replication_large_2 
Benchmark started

latencies: 69.2µs min, 5.9ms avg, 23.2ms max, 5.6ms p0.50, 12.6ms p0.95, 14.6ms p0.99
50000 total records sent, 99800 records/sec: (512.0 MB/sec), total time: 501.5ms

**Per Record E2E Latency**

|Variable| p0.00 |p0.50|p0.95 |p0.99 |p1.00 |
|--------|-------|-----|------|------|------|
|Latency |69.2µs |5.6ms|12.6ms|14.6ms|23.2ms|

**Throughput (Total Produced Bytes / Time)**

|     Variable      |   Speed    |
|-------------------|------------|
|Produced Throughput|512.0 MB/sec|

Benchmark completed
```