@fraidev 2025-03-31
> Trace shows no record were sent, even thought slice was 1MB. So there should be at least 1 record arriving. Can you add more debugging to see how many bytes were received and why decoding failed?
> 
> ```
> 2025-03-31T22:32:13.626134Z TRACE respond{socket=Socket(12) _connection=peer("127.0.0.1:64972")}:LeaderConnection{self=5002 sink=fd(12)}:update_from_leaders{sink=fd(12)}: fluvio_spu_schema::file: KfFileRecordSet encoding file slice len: 1000000
> 2025-03-31T22:32:13.649397Z DEBUG respond{socket=Socket(12) _connection=peer("127.0.0.1:64972")}:LeaderConnection{self=5002 sink=fd(12)}: fluvio_spu::replication::leader::connection: waiting for next event
> 2025-03-31T22:32:13.649391Z TRACE FollowerGroupController{leader=5001}: fluvio_spu::replication::follower::peer_api: decoding with header: RequestHeader {
>     api_key: 0,
>     api_version: 7,
>     correlation_id: 1,
>     client_id: "leader: 5001",
> }
> 2025-03-31T22:32:13.649500Z DEBUG FollowerGroupController{leader=5001}:sync_from_leader{sink=fd(11)}: fluvio_spu::replication::follower::controller::inner: update from leader replica=test-0 leader_hw=0 leader_leo=5 records=0 base_offset=-1
> 2025-03-31T22:32:13.649528Z DEBUG FollowerGroupController{leader=5001}:sync_from_leader{sink=fd(11)}: fluvio_spu::replication::follower::state: no records
> 2025-03-31T22:32:13.649540Z DEBUG FollowerGroupController{leader=5001}:sync_from_leader{sink=fd(11)}: fluvio_spu::replication::follower::controller::inner: no changes
> 2025-03-31T22:32:13.649560Z DEBUG FollowerGroupController{leader=5001}: fluvio_spu::replication::follower::controller::inner: waiting request from leader counter=1
> test replication::test::test_sync_larger_records_sync ... FAILED
> ```

If I understood well, these logs are saying that the records were not received by follower. But it's not saying about leader sending it.