@sehz 2025-04-01
Let's change in here `warn` to `debug`: https://github.com/infinyon/fluvio/blob/master/crates/fluvio-protocol/src/record/data.rs#L320 for both cases since it is expected that `RecordSet` typical will not able to process all the batches and warning is misleading